+++ extensions/tacview/extension.py
@@ -324,17 +324,29 @@ class Tacview(Extension):
         return utils.get_windows_version(os.path.join(path, 'tacview.dll'))
 
     async def install(self) -> bool:
-        def ignore_export_lua(dirname, filenames):
-            return ['Scripts/Export.lua'] if dirname == from_path else []
+        def ignore_funct(dirname, filenames) -> list[str]:
+            ignored = []
+            for item in filenames:
+                path = os.path.join(dirname, item)
+                relpath = os.path.relpath(path, from_path)
+                to_path = os.path.join(self.server.instance.home, relpath)
+                if utils.is_junction(to_path):
+                    ignored.append(item)
+                elif relpath == 'Scripts\\Export.lua':
+                    ignored.append(item)
+            return ignored
 
         if not self.get_inst_path():
             self.log.error("You need to specify an installation path for Tacview!")
             return False
         from_path = os.path.join(self.get_inst_path(), 'DCS')
-        shutil.copytree(from_path, self.server.instance.home, dirs_exist_ok=True, ignore=ignore_export_lua)
+        shutil.copytree(from_path, self.server.instance.home, dirs_exist_ok=True, ignore=ignore_funct)
         export_file = os.path.join(self.server.instance.home, 'Scripts', 'Export.lua')
-        async with aiofiles.open(export_file, mode='r', encoding='utf-8') as infile:
-            lines = await infile.readlines()
+        try:
+            async with aiofiles.open(export_file, mode='r', encoding='utf-8') as infile:
+                lines = await infile.readlines()
+        except FileNotFoundError:
+            lines = []
         if TACVIEW_EXPORT_LINE not in lines:
             lines.append(TACVIEW_EXPORT_LINE)
             async with aiofiles.open(export_file, mode='w', encoding='utf-8') as outfile:
@@ -360,7 +372,7 @@ class Tacview(Extension):
             for name in dirs:
                 dir_x = os.path.join(root, name)
                 dir_y = dir_x.replace(from_path, self.server.instance.home)
-                if os.path.exists(dir_y):
+                if os.path.exists(dir_y) and not utils.is_junction(dir_y):
                     try:
                         os.rmdir(dir_y)  # only removes empty directories
                     except OSError:
